Reorganized the directories & files
