+++ src/main.py
@@ -1,0 +1,1 @@
+# A main function to call all the api functios
